+++ engine/CHANGELOG.md
@@ -1,4 +1,9 @@
-# rc: minor bump
+# rc: micro bump
+
+- Fix pinch-to-zoom on Firefox (on Linux and Android, at least) - it seems that
+  its pointerIds sometimes just suddenly change on you.
+
+# @wwtelescope/engine 7.5.0 (2020-10-09)
 
 - Make sure crosshairs stay square when viewport size changes
 - Add TypeScript bindings to the `WWTControl.zoom()` API.